requirements.txt
mistralai
python-dotenv
ipython
